+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 0.3.4 - unreleased
 ------------------
 
-
+* Test ``analog.utils`` implementation.
 
 0.3.3 - 2014-03-10
 ------------------